+++ src/app/service/poap/OptInPOAP.ts
@@ -56,10 +56,19 @@ const OptInPOAP = async (user: User, dmChannel: DMChannel): Promise<void> => {
 					.setStyle('SECONDARY'),
 			);
 		Log.debug('user has DMs option turned off, now asking user for opt-in to get DM POAPs');
-		const message: Message = await dmChannel.send({
+		const message: Message | void = await dmChannel.send({
 			content: 'I can send you POAPs directly to you. Would you like me to do that going forward?',
 			components: [row],
+		}).catch(e => {
+			LogUtils.logError('failed to ask for opt-in', e);
+			return;
 		});
+		
+		if (message == null) {
+			Log.debug('did not sent opt-in message');
+			return;
+		}
+		
 		// 5 minute timeout
 		try {
 			await message.awaitMessageComponent({
@@ -102,7 +111,9 @@ const OptInPOAP = async (user: User, dmChannel: DMChannel): Promise<void> => {
 		Log.debug('user settings update skipped');
 	} else {
 		Log.debug(`user is opted in to dms, userId: ${user.id}`);
-		await dmChannel.send({ content: 'I will send you POAPs as soon as I get them!' });
+		await dmChannel.send({ content: 'I will send you POAPs as soon as I get them!' }).catch(e => {
+			LogUtils.logError('failed to send opt-in confirmation', e);
+		});
 	}
 };